backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==2.3.7